// src/r/types.ts
export namespace NpmNS {
  export interface IPackageJSON {
    name: string;
    version: string;
  }

  export interface IPackage {
    dirname: string;
    name: string;
    pkg: Record<string, unknown>;
  }

  export interface IRemotePackageInfo {
    name: string;
    dirname: string;
    version: string;
    path: string;
  }
}